refact: move requireAuth() to a utils
